# Fix AQI calculation
Previous AQI calculation was wrong as the reported values from the RSQA were
the AQI contributions, and not the actual measured values. The new calculation
fixes this and reports the proper AQI.

Correspondingly, the pollutant levels are back-calculated from the AQI values.
The levels aren't exact as the AQI values reported are rounded to the closest
integer.

## Minor changes

Starting to write a function to calculate the AQI value based on the EPA
formula. This is a WIP.

Bumping to version 0.2.0